split fs creator
